backend/requirements.txt
fastapi
uvicorn
pydantic>=2.0
pydantic-settings
langgraph
langchain-core
langchain-google-vertexai
google-cloud-aiplatform
python-dotenv
sqlmodel
httpx
openmeteo-requests
requests-cache
retry
retry-requests
duckduckgo-search>=6.3.0
beautifulsoup4==4.12.3
requests==2.31.0
streamlit
pandas
chromadb
sentence-transformers
openai
